core/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Matus-Abrhan <xabrhan@stuba.sk>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
pexpect = "^4.9.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
flake8-bugbear = "^24.4.26"
mypy = "^1.10.0"
pytest = "^8.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "server.run:run"

[tool.mypy]
exclude = [
	'iASTD',
]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
